Add api to handle scores
