+++ package.json
@@ -8,17 +8,16 @@
     "dev": "webpack serve --mode development",
     "build": "webpack --mode production"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Ralph-1/Endless-Runner.git"
+  "jest": {
+    "automock": false,
+    "setupFiles": [
+      "./setupJest.js",
+      "jest-canvas-mock"
+    ]
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/Ralph-1/Endless-Runner/issues"
-  },
-  "homepage": "https://github.com/Ralph-1/Endless-Runner#readme",
   "devDependencies": {
     "@babel/core": "^7.7.2",
     "@babel/preset-env": "^7.7.1",
@@ -29,6 +28,7 @@
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-stage-0": "^6.24.1",
     "clean-webpack-plugin": "^3.0.0",
+    "css-loader": "^5.2.4",
     "eslint": "^6.8.0",
     "eslint-config-airbnb-base": "^14.1.0",
     "eslint-plugin-import": "^2.20.2",
@@ -39,25 +39,23 @@
     "raw-loader": "^3.1.0",
     "sass": "^1.32.11",
     "sass-loader": "^11.0.1",
+    "style-loader": "^2.0.0",
     "stylelint": "^13.3.3",
     "stylelint-config-standard": "^20.0.0",
     "stylelint-csstree-validator": "^1.9.0",
     "stylelint-scss": "^3.17.2",
     "terser-webpack-plugin": "^2.2.1",
-    "css-loader": "^6.1.0",
-    "html-loader": "^2.1.2",
-    "style-loader": "^3.1.0",
-    "webpack": "^5.44.0",
-    "webpack-cli": "^4.7.2",
-    "webpack-dev-server": "^3.11.2",
     "url-loader": "^4.1.1",
+    "webpack": "^4.41.2",
+    "webpack-cli": "^4.6.0",
+    "webpack-dev-server": "^3.11.2",
     "webpack-merge": "^4.2.2"
   },
   "dependencies": {
     "@jest/globals": "^26.6.2",
     "clean-webpack-plugin": "^4.0.0-alpha.0",
     "html-webpack-plugin": "^5.3.1",
-    "phaser": "^3.55.2",
+    "phaser": "^3.54.0",
     "regenerator-runtime": "^0.13.7",
     "terser-webpack-plugin": "^5.1.1",
     "webpack": "^5.35.1",
