--- src/config/config.js
@@ -1,11 +1,0 @@
-import 'phaser';
-
-export default {
-  type: Phaser.AUTO,
-  parent: 'phaser-example',
-  width: 800,
-  height: 600,
-  physics: {
-    default: 'arcade'
-  }
-};
