+++ src/js/actions/destroyForm.js
@@ -1,0 +1,4 @@
+export default function fn() {
+  const form = document.querySelector('.nameForm');
+  form.remove();
+}
