+++ src/js/actions/destroyLeaderBoard.js
@@ -1,0 +1,8 @@
+export default function fn() {
+  const h1 = document.querySelector('h2');
+  const table = document.querySelector('table');
+  const credits = document.querySelector('p');
+  h1.remove();
+  table.remove();
+  credits.remove();
+}
