+++ src/js/actions/destroyTitle.js
@@ -1,0 +1,8 @@
+export default function fn() {
+  const h1 = document.querySelector('h1');
+  const div = document.querySelector('div');
+  const para = document.querySelector('p');
+  h1.remove();
+  div.remove();
+  para.remove();
+}
