+++ src/js/actions/getName.js
@@ -1,0 +1,6 @@
+const findInputValue = () => document.querySelector('input').value;
+
+export default function fn() {
+  if (findInputValue() === '') { return 'unknown'; }
+  return findInputValue();
+}
