+++ src/js/config/gameConfig.js
@@ -1,0 +1,11 @@
+// eslint-disable-next-line import/no-unresolved
+import Phaser from 'phaser';
+
+export default {
+  type: Phaser.AUTO,
+  width: 800,
+  height: 600,
+  physics: {
+    default: 'arcade',
+  },
+};
