+++ src/js/config/gameOptions.js
@@ -1,0 +1,8 @@
+export default {
+  obstacleHeight: [300, 485],
+  obstacleSpeed: [400, 500],
+  playerGravity: 900,
+  jumpForce: 500,
+  playerStartPosition: 100,
+  jumps: 2,
+};
