--- src/scenes/GameScene.js
@@ -1,11 +1,0 @@
-import Phaser from 'phaser';
-
-export default class GameScene extends Phaser.Scene {
-  constructor () {
-    super('Game');
-  }
-  
-  create () {
-    this.scene.start('Preloader')
-  }
-};
